RichText - LinkPlugin deaktivierbar machen und BubbleMenuItems zu Plugin hinzufügen #1605
