+++ .changeset/bright-onions-thank.md
@@ -1,0 +1,5 @@
+---
+"@neuvernetzung/design-system": patch
+---
+
+RichText - LinkPlugin deaktivierbar machen und BubbleMenuItems zu Plugin hinzufügen #1605
